+++ src/docu_bot/evaluation/evaluate.py
@@ -1,5 +1,8 @@
+import time
+from typing import Tuple, List
+from tqdm import tqdm
 from collections import defaultdict
-from docu_bot.constants import PROMPTS
+from datasets import Dataset
 from ragas.metrics import (
     FactualCorrectness,
     Faithfulness,
@@ -12,12 +15,14 @@ from ragas.metrics import (
 )
 from ragas.llms import LangchainLLMWrapper
 from ragas.embeddings import LangchainEmbeddingsWrapper
-from datasets import Dataset
-from langchain_openai import ChatOpenAI, OpenAIEmbeddings
+from ragas.evaluation import evaluate, EvaluationResult
+from ragas.testset.synthesizers.testset_schema import Testset
 from langchain.prompts import ChatPromptTemplate
+from langchain.retrievers import MultiVectorRetriever
 from langchain_core.runnables.passthrough import RunnablePassthrough
 from langchain_core.output_parsers.string import StrOutputParser
-from ragas.evaluation import evaluate
+from langchain_openai import ChatOpenAI, OpenAIEmbeddings
+from docu_bot.constants import PROMPTS
 from docu_bot.utils import format_docs
 
 
@@ -46,11 +51,12 @@ class Evaluator:
 
     def evaluate_configuration(
         self,
-        dataset,
-        rag_llm,
-        document_retriever,
+        dataset: Testset,
+        rag_llm: ChatOpenAI,
+        document_retriever: MultiVectorRetriever,
         prompt=PROMPTS.SYSTEM_PROMPT,
-    ):
+        use_tqdm=True,
+    ) -> Tuple[EvaluationResult, List]:
         chat_prompt = ChatPromptTemplate.from_messages(
             [("user", prompt + PROMPTS.INPUT_PROMPT)]
         )
@@ -65,8 +71,15 @@ class Evaluator:
             | StrOutputParser()
         )
 
+        time_data = []
         output_dataset = defaultdict(list)
-        for sample in dataset.samples:
+        if use_tqdm:
+            sample_iterate = tqdm(dataset.samples)
+        else:
+            sample_iterate = dataset.samples
+
+        for sample in sample_iterate:
+            gen_ts_start = time.time()
             output_dataset["user_input"].append(sample.eval_sample.user_input)
             output_dataset["reference"].append(sample.eval_sample.reference)
             output_dataset["reference_contexts"].append(
@@ -76,17 +89,21 @@ class Evaluator:
                 rag_chain.invoke(sample.eval_sample.user_input)
             )
             context = document_retriever.invoke(sample.eval_sample.user_input)
+            gen_ts_end = time.time()
             if len(context) == 0:
                 output_dataset["retrieved_contexts"].append([""])
             else:
                 output_dataset["retrieved_contexts"].append(
                     [document.page_content for document in context]
                 )
+            time_data.append(gen_ts_end - gen_ts_start)
 
         output_dataset = Dataset.from_dict(output_dataset)
-        return evaluate(
+
+        eval_result = evaluate(
             dataset=output_dataset,
             metrics=self.metrics,
             llm=self.evaluator_llm,
             embeddings=self.evaluator_embedding_model,
         )
+        return eval_result, time_data